Workaround compiler bug
The location pragma added to the documented code snippets
seems to cause the `private` modifier to be dropped.
Removing the `#<loc:FILE` part from
generate_specs_for_docs.cr on line 29 fixes the issue.
But, this would remove the ability to trace the code back to the docs.

Modified the docs for `Double.def_define_double`.
It appears that the problem it described no longer exists.